fix: align clustering engine behavior

+++ skyulf-core/skyulf/preprocessing/bucketing.py
@@ -96,8 +96,9 @@ def _polars_one_col_expr(
     cut_expr = _polars_cut_expr(col, sorted_edges, labels, include_lowest)
     target_col_name = f"{col}{output_suffix}"
     if label_format in ("ordinal", "bin_index") and not has_valid_custom_labels:
-        # Polars cut returns Categorical; cast → physical index.
-        return cut_expr.cast(pl.UInt32).alias(target_col_name)
+        # Match pandas' labels=False output dtype, including when no values are
+        # missing and pandas can retain a native int64 series.
+        return cut_expr.cast(pl.Int64).alias(target_col_name)
     return cut_expr.alias(target_col_name)
